Move the encode_decode test in a separate file
