--- src/aom.rs
@@ -1,6 +1,0 @@
-#![allow(dead_code)]
-#![allow(non_camel_case_types)]
-#![allow(non_snake_case)]
-#![allow(non_upper_case_globals)]
-
-include!(concat!(env!("OUT_DIR"), "/aom.rs"));
